Extract podcast cover transformation
